pyproject.toml
[project]
name = "ci"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.44",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-icdiff>=0.9",
    "pyyaml>=6.0.2",
]
